Makefile working

+++ tests/functional/bench.h
@@ -2,14 +2,14 @@
 #define BENCH_H
 
 // libc benchmarkings
-void bench_malloc_sparse(int N);
-void bench_malloc_bubble(int N);
-void bench_malloc_tiny1(int N);
-void bench_malloc_tiny2(int N);
-void bench_malloc_big1(int N);
-void bench_malloc_big2(int N);
-void bench_malloc_thread_stress(int n);
-void bench_malloc_thread_local(int n);
+void bench_malloc_sparse(unsigned long N);
+void bench_malloc_bubble(unsigned long N);
+void bench_malloc_tiny1(unsigned long N);
+void bench_malloc_tiny2(unsigned long N);
+void bench_malloc_big1(unsigned long N);
+void bench_malloc_big2(unsigned long N);
+void bench_malloc_thread_stress(unsigned long n);
+void bench_malloc_thread_local(unsigned long n);
 
 // my_malloc benchmarkings
 void malloc_then_free_then_malloc(void);
